Remove default favicon and rename schema file
