+++ apps/kg-explorer/src/app/digital-objects-metadata.schema.ts
